Move asset data to modules
