+++ engine/src/quoll/animation/AnimationStateLuaTable.cpp
@@ -1,6 +1,6 @@
 #include "quoll/core/Base.h"
-#include "quoll/asset/AnimatorAsset.h"
 #include "AnimationStateLuaTable.h"
+#include "AnimatorAsset.h"
 
 namespace quoll {
