+++ engine/src/quoll/asset/AssetCacheFont.cpp
@@ -1,8 +1,8 @@
 #include "quoll/core/Base.h"
+#include "quoll/text/FontAsset.h"
 #include "quoll/text/MsdfAtlas.h"
 #include "quoll/text/MsdfLoader.h"
 #include "AssetCache.h"
-#include "FontAsset.h"
 #include <freetype/freetype.h>
 #include <ft2build.h>
