+++ engine/src/quoll/scene/EnvironmentAsset.h
