+++ engine/src/quoll/text/FontAsset.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "quoll/asset/Asset.h"
+#include "quoll/rhi/RenderHandle.h"
+
 namespace quoll {
 
 struct FontGlyph {
