+++ engine/src/quoll/text/MsdfLoader.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "quoll/asset/AssetData.h"
-#include "quoll/asset/FontAsset.h"
 #include "quoll/asset/Result.h"
+#include "FontAsset.h"
 
 namespace quoll {
