+++ engine/tests/quoll-tests/animation/KeyframeInterpolator.test.cpp
@@ -1,6 +1,6 @@
 #include "quoll/core/Base.h"
+#include "quoll/animation/AnimationAsset.h"
 #include "quoll/animation/KeyframeInterpolator.h"
-#include "quoll/asset/AnimationAsset.h"
 #include "quoll-tests/Testing.h"
 
 using SequenceTarget = quoll::KeyframeSequenceAssetTarget;
